refactor: remove legacy bank write operations and update imports
