--- app/src/composables/bank/writes-new.ts
@@ -1,3 +1,0 @@
-// Legacy import - now uses the generic contract writes system
-// This file is kept for backward compatibility
-export { useBankWrites } from './useBankWrites'
